pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "python-docx>=1.2.0",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
    "tiktoken>=0.11.0",
]
